Suit.java:32-34: Move this interface to the Core module....
#18 More questions.
